refactor: split file persistence from repository
Moved everything not file persistence related to repository class.
